buggybase2/buggybase2/requirements.txt
Flask-Bootstrap4==4.0.2
Flask==1.1.2
uWSGI==2.0.19.1